Use a custom Less fork with slight modification
This will use a slight modification to how Less `@import (optional)`
works. As described in the Less docs, optional imports silently discard
errors with missing files. Unfortunately, the implementation also
discards *all other errors*. The change in the commit below allows Less
to continue if there is a file error, but not if there is a parsing
error. This makes it far easier to discover underlying issues with
optionally included files, which is pretty much all of the SUI theme.

Commit is at:
https://github.com/agjohnson/less.js/commit/64179a24eef7308dcfb730bf0fcc3753e60ba026

This also runs npm update, which grabbed an updated Babel and friends.

You'll need to run `npm ci` locally after this is merged.