Add automatic linting check
This uses prettier and all of the opinionated defaults from prettier.
I'm not always a fan of some of the decisions, but it allows us to not
think about code style.

This enables CircleCI checks on branches and pull requests. It's also
usable locally.

Also drop fomantic-ui dependency to avoid building sui and pin node for CircleCI