Overhaul homepage
This is a big update, unraveling quite a bit of history trying to revamp
the content and graphic elements. It replaces the homepage above the
fold graphics with easier to maintain graphics in place of third party
custom graphics. The only custom graphic is the main hero graphic, used
to fill the space a bit more and to tie things together.

Graphic changes are:

* The main hero image was replaced. It was an unfinished stock art
  screengrab.
* The 3 feature section graphics were replaced with colored FA duotone
  graphics.

Overall, this provides a more polished feel, matches the graphics in the
rest of the site without generating more custom glyphs, and added color
and detail is nicer than the previous graphics.

Structure changes are:

* Modal for sign up selection, triggered by the hero CTA
* 6 grid feature overview on the front page to add more content
* Blog articles section was dropped to the bottom of the page and
  replaced with "Resources" instead.

Content changes dictated a lot of these updates.

Closes #127
Closes #42
Closes #53
Closes #41
Closes #32
Closes #122
Refs #128