Rebase about page commits
This is the old "product" page, which sounded better as "about" instead
in the nav.

I addressed #123, but also focused quite a bit on the content. We were
missing a lot of features, but still using a lot of words to describe
the features we were covering. I focused on clearing up language and
focusing on the essentials, and included a few product screenshots,
where we haven't had any yet.

Still WIP at the moment, there are placeholders sprinkled through. Draft
PR for now just to get some of the work off my laptop, it's not review
ready.