Rework product/about page (#147)
* Rebase about page commits

This is the old "product" page, which sounded better as "about" instead
in the nav.

I addressed #123, but also focused quite a bit on the content. We were
missing a lot of features, but still using a lot of words to describe
the features we were covering. I focused on clearing up language and
focusing on the essentials, and included a few product screenshots,
where we haven't had any yet.

Still WIP at the moment, there are placeholders sprinkled through. Draft
PR for now just to get some of the work off my laptop, it's not review
ready.

* Moving around some feature blocks and finishing header content and images

Images were mostly placeholder, I put a bit more thought in this pass. I
dropped an image for now, and added a UI screenshot I'm not super jazzed
with, but we can update these as we go.

* Update last few blocks of about page content

* Add homepage modal for site selection to about page

* Fix button attr

* Update webhook and notification copy/icon

Tweaked copy a bit on both sections. It felt a little bit unfinished
actually, I might have missed coming back to these.

* Apply suggestions from code review

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>

* Update about page lead heading

* Update host features block

* Drop down to 3 feature blocks. 4th block on the first line caused a
  misalignment with the header and the icons below.
* Improve copy in the left side and put a temporary block of content
  about version privacy in the right

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>