Page: docs-as-code
Initial work to talk about "Docs-as-code concept" and how it's integrated with
Read the Docs getting the main features by default thanks to the
`readthedocs-client.js`.

The idea is to expose all these main features as
https://github.com/readthedocs/readthedocs-client does by talking about
docs-as-code and "all the doctools".

At the bottom there should be a copy&paste examples with one pretty known
doctool different than Sphinx and MkDocs (maybe Docusaurus).