Some fixes for the front page header changes
Some leftovers from #211. Fixes capitalization to match our other
headers and puts the focus not on time but on our service.

I've had this feedback on other similar CTAs. Focusing on
minutes/seconds is saying our value is that we save time. That's not
wrong, but we don't exactly aim to save time, but rather our value is
that we manage all of the hurdles with infrastructure when building and
hosting authed/searchable/etc docs.