Improve buttons in sign up modal
I found that the signup modal doesn't link to the "Choosing a platform" page,
which I think it's important in that context. The modal shows small differences
between the two sites, but I felt there was some context missing there. I added
a link to this page so users can dig deeper if they want to.

Besides, I changed the copy from "Learn more" to "Explore all the features",
that communicates exactly what that button is for.

Finally, I added icons to these buttons because they make everything to look better.