Page: docs-as-code (#194)
* Page: docs-as-code

Initial work to talk about "Docs-as-code concept" and how it's integrated with
Read the Docs getting the main features by default thanks to the
`readthedocs-client.js`.

The idea is to expose all these main features as
https://github.com/readthedocs/readthedocs-client does by talking about
docs-as-code and "all the doctools".

At the bottom there should be a copy&paste examples with one pretty known
doctool different than Sphinx and MkDocs (maybe Docusaurus).

* Update the page with the initial structure and idea

* Add flyout section

* Add this page to the footer

* Use images from Docusaurus doctool

It matches the example shown below and it looks better also.

* Small copy changes

* Describe the simple steps for "docs as code"

* More copy changes

* Minor change

* Update flyout with languages

* Lint issues

---------

Co-authored-by: Eric Holscher <eric@ericholscher.com>