Use filename as slug source
By default, the title of the article is used as the slug,
if the title is changed, the slug will be updated,
and the URL of the article will be changed, leading to a 404.

Our previous blog used the filename as the slug, I like that.

If we agree, I can set slug explicitly for the articles
that are already published or maybe create a redirect.