Convert top navigation to multiple dropdown menus (#221)
* Add example navigation menu for dropdown product menu

This is for:

- #176

But this is still just proof of concept level so far, the mobile menu is
a bit broken by this and the actual structure needs shaping up still.

* Polish up navigation menus

This drops the concept of highlighting the navigation menu item on the
page slug, but adds two menus we can put content under to make content
discoverable.

This now works on desktop and mobile.

* Put pricing link in top level topnav too

I left the submenu option under product too, it looks nice there. The
size of the menu was reduced to match the beta dashboard, so elements
could fit.