Pricing: mention custom domain limitations
We agreed on:

- 2 domains per project on community
- 5 domains per organization on Advanced
- 15 domains per organization on Pro

This is the first step: update our pricing page for new customer.
One of the next steps involves enforcing this in our application by adding Read
the Docs features to the plans and also showing the domains left in the admin.

Related https://github.com/readthedocs/readthedocs-corporate/issues/1808