Landing: day to day usage
Example of "How to use Read the Docs in your daily work" after reading the
onboarding book on Book Club.

The idea of this page is to show _exactly_ how customers will use Read the Docs
on their day to day work to make them reach that "aha" moment when they realize
this workflow is a lot easier than the one they are currently following.

Goals:

- explain how they will use Read the Docs daily
- mention docs as code workflow
- explain write, push, review and deploy stages
- show images representing those exact stages
- mention we don't provide an editor

Page based on https://about.readthedocs.com/docs-as-code/ structure.